ref(rust): Ensure none keys are properly stripped from broker config
Ensure broker config is always properly deserialized. Previously this
was done in the consumer code, but we should move it to the config, so
it will also work in other scenarios where we need broker config - such
as the replacements producer and the commit log producer.
